move list.hpp up
