+++ src/bsoncxx/test/bson_builder.cpp
@@ -15,7 +15,7 @@
 #include <bsoncxx/builder/basic/array.hpp>
 #include <bsoncxx/builder/basic/document.hpp>
 #include <bsoncxx/builder/core.hpp>
-#include <bsoncxx/builder/list/list.hpp>
+#include <bsoncxx/builder/list.hpp>
 #include <bsoncxx/builder/stream/array.hpp>
 #include <bsoncxx/builder/stream/document.hpp>
 #include <bsoncxx/exception/exception.hpp>
